backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
motor>=3.3.0
pymongo>=4.6.0
beanie>=1.24.0
python-dotenv>=1.0.0
pillow>=10.0.0
aiofiles>=23.0.0